Change LinkI2C to permit operation to throw on read or write. Use to implement ping test to check on a device.
For non-ping consumers of read/write in devices, use try! (This is the same behavior as previous versions where the non-throwing
implementations would fatalError on an unexpected NACK.)

Add PresenceQuery enum for ping strategy. Make that an option with I2CTraits.

PlatformDeviceBroker:
- make singleton
- add findI2C variant that sets up and pings link based on desired class' I2CTraits
- with multiple searches for a node, do not create multiple Ftdi instances: reuse the one found instance

ftdi-synchronous-serial: upgrade required version to 0.5.3; remove sleeps that seemed to help pre-0.5.3 NACK problem.